refactor: remove backend deps from form options
