+++ sites/partners/src/users/FormUserManage.tsx
@@ -14,14 +14,13 @@ import {
   AuthContext,
   AppearanceStyleType,
   AppearanceBorderType,
-  RoleOption,
-  roleKeys,
   emailRegex,
   setSiteAlertMessage,
   Tag,
   AppearanceSizeType,
   Modal,
 } from "@bloom-housing/ui-components"
+import { RoleOption, roleKeys } from "@bloom-housing/shared-helpers"
 import { Listing, User, UserRolesCreate } from "@bloom-housing/backend-core/types"
 import router from "next/router"
