+++ ui-components/src/helpers/formOptions.tsx
@@ -1,20 +1,11 @@
 import * as React from "react"
 import { t } from "./translator"
-import { Language } from "@bloom-housing/backend-core/types"
 import { SelectOption } from "../forms/Select"
 
 export interface FormOptionsProps {
   options: (string | SelectOption)[]
   keyPrefix?: string
 }
-
-export const applicationLanguageKeys = [Language.en, Language.es, Language.zh, Language.vi]
-
-export enum RoleOption {
-  Administrator = "administrator",
-  Partner = "partner",
-}
-export const roleKeys = Object.values(RoleOption)
 
 export const numberOptions = (end: number, start = 1): SelectOption[] => {
   const nums = []
